Upgrade osm2orc + upload direct to S3
s3a was previously being used, but it was buffering files in full before
uploading. osm2orc-0.1.1 is required to load conf/core-site.xml (as
conf/ was not previously in the CLASSPATH and couldn't be easily
modified at runtime).

To avoid running out of heap space when buffering uploads, the max heap
size has been increased to 6GB.